Use PID1 as initi process
Using a process htat correctly handles signals is important to avoid
zombie processes as well as slow shutdown times.

More info is available at:
https://www.fpcomplete.com/blog/2016/10/docker-demons-pid1-orphans-zombies-signals

This patch also adds the ability to change the user within the
container, which provides added security in case of a Docker jail break.
And it avoids permission issues on files in directories mounted from the
host.