Lookup album image from Discogs if possible if attaching from file fails
When syncing the media library, album art is looked up from the file, and if not found, is left blank even when the art can be found on discogs.

This change automatically queues a job to attempt to attach the album art from discogs if not found when attaching from the file - and discogs lookup is available.