Let title available on all pages
Because on Turbo Android, if a page doesn't have a title, it will set the current page URL as the view title. It seems to be a default behavior on android web view. So to avoid this, I need to enable title on all page no matter the page is on native web view or not.