Fix the modal dialog can not dismiss automatically
Because black candy use turbo custom rendering to avoid reloading the permanent aside player element. Which means the content that not in app container won't update in normal web page navigation. So we need to put the dialog inside of the app container to avoid dialog can not dismiss automatically after form submitted in dialog.